re organize js files
