+++ components/02-components/filters/range-limited.html
@@ -46,7 +46,7 @@
           <span class="u-visually-hidden">Search</span>
         </button>
       </div>
-      <div class="date-range__grid js-date-grid" style="display: block">
+      <div class="date-range__grid js-date-grid">
         <div class="date-range__row">
           <div class="date-range__year">2017</div>
           <ul data-year="2017" class="date-range__months">
